ai-service/requirements.txt
google-generativeai==0.8.3
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pillow==10.4.0
python-multipart==0.0.6
PyPDF2==3.0.1
aiohttp==3.9.1
